requirements.txt
# Core web & API
fastapi
uvicorn[standard]

# (Debug) JSONL-based pipeline uses lexical retrieval.
# (Next step) Chroma + embeddings:

chromadb
# sentence-transformers   # deferred until we decide embedding strategy
# langchain

# Utilities
httpx
python-dotenv
pydantic

# Progress / UX
tqdm

# Doc Scanning

pypdf
openpyxl
python-docx
python-pptx


# Development

pre-commit
pytest
ruff
setuptools
wheel
